Fix ServiceWorkerAutoPreload response handling
This CL fixes the issue of ServiceWorkerAutoPreload.

Before this, when the ServiceWorkerAutoPreload is used, the navigation
response is not delivered to the renderer correctly. That results in
showing the blank page. This issue was introduced by crrev.com/c/7134458
which was landed in M143, and this CL will add the fix for that.

In crrev.com/c/7134458, the new argument was added to
ServiceWorkerRaceNetworkRequestURLLoaderClient constructior, which is
clone_completed_for_fetch_handler_callback. This was introduced to
invalidate the SWMainResourceLoader so that the SWMainResourceLoader is
destructed after the fetch handler completion when
race-network-and-fetch-handler is specified via the Static Routing API.
However, this introduced the task scheduling issue in SWAutoPreload, as
the SWAutoPreload uses the same URLLoaderClient, and it accidentally
deleted the object while the response processing is not completed yet.

This CL fixes the issue by passing a no-op callback to
clone_completed_for_fetch_handler_callback when the URLLoader is created
for the SWAutoPreload.

Also, this CL explicitly enable/disable `ProcessHtmlDataImmediately` in
tests. As `ProcessHtmlDataImmediately` changes the scheduling of HTML
parsing in the renderer side and that could obfuscate the issue, we test
in both feature state.

Please see crbug.com/468821930#comment6 for more details.

Change-Id: Iadb8092bf2a2e0bbb7d8b2b5c07c9652fb9d3531
Bug: 469525825, 469527486, 468821930
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7261754
Reviewed-by: Takashi Nakayama <tnak@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1559807}