Simplify TextCodecUtf8::HandlePartialSequence()
Fold the TextCodecUtf8::HandleError() helper into the only caller. With
this done we can merge the various branches in HandlePartialSequence()
(the UChar variant) such that it has the same structure as the main
decoding loop.

Merge the replacement character emit with the normal code point emit.
Also do this in the main (UTF-16) decode loop.

Bug: 469372291
Change-Id: I2c139569e5b2699e09a63739c03d6ea8fbe6d906
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7265208
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1559826}