Ensure ChromeAndroidTask outlives RootUiCoordinator
While working on crrev.com/c/7237655, we realized that ChromeAndroidTask
(Java) and its associated AndroidBrowserWindow (C++) are destructed
before RootUiCoordinator. This is problematic because RootUiCoordinator
owns ExtensionActionsBridge (indirectly via ToolbarManager) and the
patch will modify the bridge to borrow AndroidBrowserWindow for its
lifetime, resulting in a dangling pointer on its destruction.

This patch solves the problem by fixing the order of destructions in
AsyncInitializationActivity#onDestroy(). It also adds an integration
test to ensure the correct destruction order.

The patch also updates MultiInstanceManagerApi31 for the new destruction
order to avoid memory leak on API 31+.

Bug: 448199168
Change-Id: I3cbfbf9975657a146a5227aabfbdecc81c7ac5c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7254915
Reviewed-by: Sirisha Kavuluru <skavuluru@google.com>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Linyu He <linyuh@google.com>
Cr-Commit-Position: refs/heads/main@{#1559838}