[contextual-tasks] Update Omnibox Placeholder text
In LocationBarModelDelegate, ShouldDisplayURL() is updated to prevent
displaying the URL for chrome://contextual-tasks, similar to the NTP.

In OmniboxViewViews, the placeholder text is set to the Contextual Tasks
page title. A check in InstallPlaceholderText() verifies if the
currently visible WebContents is the Contextual Tasks and sets the page
title as the placeholder text. OmniboxViewViews starts observing the
visible WebContents for DidFinishNavigation() and TitleWasSet() events
to call InstallPlaceholderText(). These ensure the placeholder text
remains accurate on navigation events including back/forward between the
NTP and the Contextual Tasks.

Bug: 459839049
Change-Id: If631b3a3c17a0eab49c7aeec7125cc30389fa79c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7233975
Reviewed-by: Aviv Kiss <avivkiss@google.com>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1559902}