Use add mouse-release event to vertical tab strip selection tests
The test was only using ClickMouse without flags, and was instead
invoking the MousePressed handler on views directly for events with
flags. This causes issues for dragging, since a mouse-up event is
never received.

This CL updates the tests to execute perform the MouseReleased handler.

Ideally, we could use ClickMouse with modifiers, but sending mouse events with modifiers is not compatible with ChromeOS (see http://http://crbug.com/469431242#comment2).

Change-Id: Iad71d5449f0bc9a830ead4ac2403947b56e078db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7266044
Commit-Queue: Kaan Alsan <alsan@chromium.org>
Reviewed-by: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1559932}