Report render blocking behavior in Network.requestWillBeSent.
Adds an experimental RenderBlockingBehavior enum to the Network domain and includes an optional renderBlockingBehavior field in the Network.requestWillBeSent event. The InspectorNetworkAgent is updated to populate this field based on the resource's render blocking status. A new web test is added to verify the reported behavior for different script and stylesheet loading scenarios.

Bug: 41169154
Change-Id: I78737bdd15e08c77f6169037eba85be07b54be81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7269488
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1559935}