[iOS] Fix bottom corners for NTP animation
This CL updates the NTP animation to ensure that only the top corners of
the new NTP are rounded during the expansion animation.

This bug only seems to repro on certain builds (I've repro'd on Canary
but not local official build). To verify the fix on local build, I
artificially reduced the view height to expose the bottom edge and
confirmed it now remains square. Will verify again once it lands in
Canary.

Before:
https://screencast.googleplex.com/cast/NTc2OTUwNDg4ODk3OTQ1NnxiZjYwZmRhNC0zMw
After:
https://screencast.googleplex.com/cast/NTI2NDQ5ODY5MTY3MDAxNnwyYzQ3ZWQ4Ni03ZA

Bug: 459752030
Change-Id: I1ebaf6964e261310a426db4dfae7248928309adb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7266824
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Hira Mahmood <hiramahmood@google.com>
Cr-Commit-Position: refs/heads/main@{#1559960}