[devtools] Remove IPH experiment (for AI assistance in DevTools)
This removes the new badges related to the IPH experiment for the
AI assistance. It also removes the features that were tied to the
experiment that are no longer needed since we've launched.

The missing feature name will be handled correctly in all front_ends,
as the `featureName` is set as optional in the front-end, requiring
clients to check if it is actually available before using it.

Bug: 448241956
Change-Id: I644365c15a4cea480825aaebe5be63d08b643c6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7208440
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1559998}