Strip Insecure Schemes From DNS Hosts When Evaluating Override Rules
Proxy override rules will make use of DNS resolution for evaluation DNS
conditions (e.g. to test whether the client is inside or outside some
specific network). This is similar to the various functions available in
PAC scripts.

When keeping the DNS query type as unspecified with an HTTP hostname,
the DNS resolution was failing and asking to try again with HTTPS.

Changed to relying on HostPortPair for insecure schemes, and keeping
SchemeHostPort flow for secure schemes. This leads to DNS resolutions of
type A and AAAA for insecure schemes, and HTTPS type on top of those for
secure schemes.

Fixed: 469064221
Change-Id: I8fc69701af389fb0b6f0f486cad50a075c0684ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7260797
Auto-Submit: Sébastien Lalancette <seblalancette@chromium.org>
Reviewed-by: Yann Dago <ydago@chromium.org>
Commit-Queue: Sébastien Lalancette <seblalancette@chromium.org>
Reviewed-by: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560047}