IDB: refactor BackingStore API for multi-phase commit/blob writing
Previously, the contract involved passing a callback which is either
invoked synchronously with value A or invoked asynchronously with value
B. From the callsite, it's not easy to see what happens, and there's
no need to handle "synchronous with value B" or vice versa.

Now, the contract involves returning a value to indicate whether there
is async work to be done. The passed callback is only ever invoked in
the case where there is async work.

One behavioral change that comes of this is that the histogram
IndexedDB.BackingStore.CommitPhaseOne only logs outcomes for
CommitPhaseOne, instead of sometimes also including outcomes for
CommitPhaseTwo. Since the histogram was added recently, there
shouldn't be a need to rename it.

Bug: 418962704
Change-Id: Iaf9f046b2e5d580ef215ceed0bba2311a2146a07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7266010
Commit-Queue: Evan Stade <evanstade@microsoft.com>
Reviewed-by: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Reviewed-by: Evan Stade <evanstade@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1560064}