[GlicMI] Support closing all embedders for real.
There are checks in multiple places (in glic and side panel code) to exit early from Close() flow if side panel is not showing. This CL changes glic code to also call Close() on backgrounded tabs. For backgrounded tabs we will now reset active side panel entry which will result in not longer showing a side panel for that tab when it gets foregrounded.

Bug: 468325375
Change-Id: I4e560ba70c499a25dc485da96b145208215f5f79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7256882
Commit-Queue: Basia Zimirska <basiaz@google.com>
Reviewed-by: Bryant Chandler <bryantchandler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560069}