[Masonry] dense-packing: include open end of tracks when finding eligible track opening
Previously, when searching for an eligible track opening for a
multi-spanning item in dense-packing, we would only consider existing
track openings in tracks, but we did not build our track openings across
the open end of tracks.

This change builds track openings across open end of tracks as needed in
order to find the highest available track opening for multi-span items.

To account for the open endings of tracks when searching for an eligible
track opening, we:

1a) Keep track of the open end of tracks as a `TrackOpening` in
`track_collection_openings_`, with the opening starting at the track's
current running position and ending at `LayoutUnit::Max()`.

1b) Because of 1a), we can remove the `running_positions_` vector and
query for the running position of a track using the `start_position` of
the last item in `track_collection_openings_`.

2) Set `highest_eligible_track_opening` in
`GetEligibleTrackOpeningAndUpdateMasonryItemSpan` to represent the
auto-placed location of the item, so that we don't pack items into track
openings lower than the auto-placed location. We also pass in the
starting position of the auto-placed item as
`auto_placement_stacking_axis_offset` in
`GetEligibleTrackOpeningAndUpdateMasonryItemSpan`.

As decided in https://github.com/web-platform-tests/wpt/pull/56435, we
remove the dense-packing specification in
`grid-lanes-grid-placement-named-lines-002.html` and fix the references
to no longer use dense-packing as well. We create a dense-packing
version of that test,
`grid-lanes-grid-placement-named-lines-dense-packing-001.html`, with
comments added to help with interpreting the test.

The newly added test `row-reverse-dense-packing-multi-span-001.html` is
currently failing because we don't compare the TrackOpening's track to
the auto-placed location's track if they have the same running position.

Lastly, the `UpdateRunningPositionsForSpan` unit tests were removed, as
the `UpdateRunningPositionsForSpan` method is pretty thoroughly tested
by the web tests.

Bug: 343257585
Change-Id: I961698e0e16118a2aae07ec57900b56b918332c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7204663
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Celeste Pan <celestepan@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1560074}