[webgpu] Don't treat invalid texture usage as a decoding error
These can be handled along with all of the other checks for valid
texture usages, rather than in a special place that treats it as a
decoding error.

Validation of the "external" usage needs to be moved to fix the
web-exposed API; validation of the internal_usage doesn't really have to
be moved but is moved for consistency.

This fixes test cases I'm adding in
https://github.com/gpuweb/cts/pull/4527 and should allow more of the CTS
to run through in a single webpage without "crashing".

Fixed: 467442134
Change-Id: I4fd1b6894a27e1c6d3a992367a120c5ee134c512
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7250051
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560115}