Fix InitFeatureListEarlyTest
In is_chrome_branded = true builds, CachedFlag's defaultValueInTests
is not applied, even if we are running tests. This means that the
LoadNativeEarly CachedFlag has a value of false, which means that
we are not initializing the feature list early.

Therefore, this CL skips the InitFeatureListEarlyTest when
is_chrome_branded = true.

Note that we cannot use @EnableFeatures annotation to set the
LoadNativeEarly CachedFlag to true, because we are retrieving flag
values too early in startup, before the annotations are applied.

Bug: 469673531
Change-Id: I6284996ee4d3155fa447927ee2b968435f4fd96b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7270624
Commit-Queue: Martin Kong <martinkong@google.com>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560143}