[DCSI] Enforce single planar format in ProduceMemory().
Currently, SharedMemoryImageBacking::ProduceMemory() creates a
MemoryImageRepresentationImpl regardless of the format, which causes a
lazy failure later in BeginReadAccess() if the format is multiplanar.

This CL adds an explicit check to fail early in ProduceMemory() if the
format is not single-planar. This is a prerequisite for the upcoming
CompoundImageBacking architecture, where CompoundImageBacking will wrap
one or more underlying backings and be primary backing itself.

Bug: 448962784
Change-Id: I2c9bd13b2ca6a9728d85694ad0beaa1d051e845a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7269356
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Vikas Soni <vikassoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560151}