[Settings Search] Let Sync pref open from search results
Sync preference in main settings either opens a fragment via
SettingsNavigation or displays a toast if it is disabled by
administration policy. This CL makes it work in the same way when
opened from search results.

For this to work, SettingsNavigation#startSettings is given an extra
parameter (EXTRA_FRAGMENT_TAG) which is given to a fragment transition
from search to result state. This allows the fragment opened via
SettingsNavigation to behave as expected for back action/tap on title.

Bug: 469676538
Change-Id: Ic84cac90a195bbfebf6734727b3367b8aeec52a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7268507
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Moe Adel <adelm@google.com>
Cr-Commit-Position: refs/heads/main@{#1560153}