[ios] Fix EG loadMinimalAppUI on ipad-device bot
This changes loadMinimalAppUI to be an instance method, so that it can
return different values depending on the test method that is being
executed. It also updates the NTPHomeTestCase to use this, rather than
setting the cmd line param directly which caused ChromeTestCase to not
know that it as using the minimal app ui mode.

Fixed: 468067115
Change-Id: Ie3d7e53e40712518a545f41302f079fa6b439fdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7269964
Commit-Queue: Scott Yoder <scottyoder@google.com>
Reviewed-by: Ernesto Izquierdo Clua <eic@google.com>
Owners-Override: Ernesto Izquierdo Clua <eic@google.com>
Cr-Commit-Position: refs/heads/main@{#1560200}