[Extensions] Pull browser finding / creation into TabsCreateFunction
Instead of relying on OpenTabHelper's implementation of finding or
creating a browser, hoist it into TabsCreateFunction. This is
necessary to make it easier to support this process on the experimental
desktop android platform, where any creation of a browser window will be
asynchronous.

Unfortunately, this does result in some duplicate code. This was cleaner
overall than trying to split this into multiple helper methods, given
relatively small scope.

Bug: 430344931
Change-Id: I1a4ad06e2fd25e838c75bf285d7fa15d522af3ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7264456
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560218}